WSTEAM1-1068: Add csp next
