+++ ws-nextjs-app/utilities/cspHeaderResponse/index.test.ts
@@ -1,0 +1,56 @@
+/**
+ * @jest-environment node
+ */
+import { NextRequest } from 'next/server';
+import cspHeaderResponse from '.';
+
+const policies = [
+  'base-uri',
+  'form-action',
+  'frame-ancestors',
+  'object-src',
+  'default-src',
+  'child-src',
+  'connect-src',
+  'font-src',
+  'frame-src',
+  'img-src',
+  'script-src',
+  'style-src',
+  'media-src',
+  'worker-src',
+  'report-to',
+  'upgrade-insecure-requests',
+];
+
+describe('cspHeaderResponse', () => {
+  it.each(policies)('should set %s in the request CSP', policy => {
+    const response = cspHeaderResponse({
+      isAmp: false,
+      isLive: false,
+      reportOnlyOnLive: false,
+      service: 'mundo',
+      request: {} as NextRequest,
+    });
+
+    const requestCsp = response.headers.get(
+      'x-middleware-request-content-security-policy',
+    );
+
+    expect(requestCsp?.includes(policy)).toBe(true);
+  });
+
+  it.each(policies)('should set %s in the response CSP', policy => {
+    const response = cspHeaderResponse({
+      isAmp: false,
+      isLive: false,
+      reportOnlyOnLive: false,
+      service: 'mundo',
+      request: {} as NextRequest,
+    });
+
+    const requestCsp = response.headers.get('content-security-policy');
+
+    expect(requestCsp?.includes(policy)).toBe(true);
+  });
+});
